# apps/web/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/activities": {
      "get": {
        "operationId": "ActivitiesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "หมายเลขหน้า (เริ่มที่ 1)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "จำนวนรายการต่อหน้า",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "คำค้นหา (contains, ไม่สนตัวพิมพ์เล็ก-ใหญ่)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "ฟิลด์ที่ใช้เรียงลำดับ (whitelist กำหนดที่แต่ละ endpoint)",
            "schema": {
              "default": "created_at",
              "type": "string",
              "enum": [
                "created_at",
                "updated_at",
                "name",
                "start_at",
                "status"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PUBLISHED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "createdById",
            "required": false,
            "in": "query",
            "description": "กรองตามผู้สร้าง (id) — มีความหมายเต็มที่กับ Admin",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ช่วงเวลาเริ่มกิจกรรม (เริ่มต้น)",
            "schema": {
              "example": "2026-09-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ช่วงเวลาเริ่มกิจกรรม (สิ้นสุด)",
            "schema": {
              "example": "2026-09-30T23:59:59.000Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List activities (paginated; own activities for organizers, all for admins — spec §29)",
        "tags": [
          "activities"
        ]
      },
      "post": {
        "operationId": "ActivitiesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetailDto"
                }
              }
            }
          }
        },
        "summary": "Create a DRAFT activity (spec §11)",
        "tags": [
          "activities"
        ]
      }
    },
    "/api/v1/activities/{id}": {
      "get": {
        "operationId": "ActivitiesController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetailDto"
                }
              }
            }
          }
        },
        "summary": "Get an activity with its linked beacons",
        "tags": [
          "activities"
        ]
      },
      "patch": {
        "operationId": "ActivitiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetailDto"
                }
              }
            }
          }
        },
        "summary": "Update an activity (status only moves via publish/cancel — spec §35)",
        "tags": [
          "activities"
        ]
      }
    },
...
    "/api/v1/activities/{id}/attendances": {
      "get": {
        "operationId": "AttendanceController_listByActivity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "หมายเลขหน้า (เริ่มที่ 1)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "จำนวนรายการต่อหน้า",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "คำค้นหา (contains, ไม่สนตัวพิมพ์เล็ก-ใหญ่)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "ฟิลด์ที่ใช้เรียงลำดับ (whitelist กำหนดที่แต่ละ endpoint)",
            "schema": {
              "default": "check_in_at",
              "type": "string",
              "enum": [
                "check_in_at",
                "status",
                "method",
                "created_at"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PRESENT",
                "LATE",
                "EXCUSED"
              ]
            }
          },
          {
            "name": "method",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "BEACON",
                "MANUAL"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttendanceRowDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List an activity's attendance rows (search by student code/name — spec §25)",
        "tags": [
          "attendance"
        ]
      }
    },
...
    "/api/v1/me/attendances": {
      "get": {
        "operationId": "MeController_attendances",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "หมายเลขหน้า (เริ่มที่ 1)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "จำนวนรายการต่อหน้า",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "คำค้นหา (contains, ไม่สนตัวพิมพ์เล็ก-ใหญ่)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "ฟิลด์ที่ใช้เรียงลำดับ (whitelist กำหนดที่แต่ละ endpoint)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttendanceItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Caller's own attendance history, newest first (created by the attendance engine (spec §10))",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/activities": {
      "get": {
        "operationId": "MeController_activities",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "หมายเลขหน้า (เริ่มที่ 1)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "จำนวนรายการต่อหน้า",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "คำค้นหา (contains, ไม่สนตัวพิมพ์เล็ก-ใหญ่)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "ฟิลด์ที่ใช้เรียงลำดับ (whitelist กำหนดที่แต่ละ endpoint)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyActivityItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "PUBLISHED activities for the LIFF activities page (spec §35), newest start first",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/activities/{id}": {
      "get": {
        "operationId": "MeController_activity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyActivityItemDto"
                }
              }
            }
          }
        },
        "summary": "One PUBLISHED activity + the caller’s own attendance (LIFF detail view)",
        "tags": [
          "me"
        ]
      }
    }
  },
  "info": {
    "title": "LINE Beacon Attendance API",
    "description": "REST API for the LINE Beacon Attendance Management System",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BeaconResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "hwid": {
            "type": "string",
            "example": "32af519e88"
          },
          "name": {
            "type": "string",
            "example": "CS Room 101"
          },
          "location": {
            "type": "object",
            "nullable": true,
            "example": "CS101"
          },
          "description": {
            "type": "object",
            "nullable": true,
            "example": "บีคอนห้อง CS101"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "MAINTENANCE"
            ]
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-22T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-08-22T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "hwid",
          "name",
          "location",
          "description",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ActivityCreatorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "example": "somorganizer"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "ActivityResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Orientation 2026"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "location": {
            "type": "object",
            "nullable": true,
            "example": "ห้องประชุมใหญ่"
          },
          "startAt": {
            "type": "string",
            "example": "2026-09-01T13:00:00.000Z"
          },
          "endAt": {
            "type": "string",
            "example": "2026-09-01T16:00:00.000Z"
          },
          "checkinOpenAt": {
            "type": "string",
            "example": "2026-09-01T12:30:00.000Z"
          },
          "lateAt": {
            "type": "string",
            "example": "2026-09-01T14:00:00.000Z"
          },
          "checkinCloseAt": {
            "type": "string",
            "example": "2026-09-01T15:00:00.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "CANCELLED"
            ]
          },
          "timeState": {
            "type": "string",
            "enum": [
              "UPCOMING",
              "CHECKIN_OPEN",
              "ONGOING",
              "COMPLETED"
            ]
          },
          "createdBy": {
            "type": "string"
          },
          "creator": {
            "$ref": "#/components/schemas/ActivityCreatorDto"
          },
          "beaconCount": {
            "type": "number",
            "description": "จำนวนบีคอนที่ลิงก์อยู่"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-22T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-08-22T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "location",
          "startAt",
          "endAt",
          "checkinOpenAt",
          "lateAt",
          "checkinCloseAt",
          "status",
          "timeState",
          "createdBy",
          "creator",
          "beaconCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "AttendanceSummaryDto": {
        "type": "object",
        "properties": {
          "totalStudents": {
            "type": "number",
            "description": "นักศึกษา ACTIVE ทั้งหมด (spec §24)"
          },
          "present": {
            "type": "number"
          },
          "late": {
            "type": "number"
          },
          "excused": {
            "type": "number"
          },
          "absent": {
            "type": "number",
            "description": "คำนวณ: Total − Present − Late − Excused (ไม่ persist, spec §44)"
          }
        },
        "required": [
          "totalStudents",
          "present",
          "late",
          "excused",
          "absent"
        ]
      },
      "ActivityDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Orientation 2026"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "location": {
            "type": "object",
            "nullable": true,
            "example": "ห้องประชุมใหญ่"
          },
          "startAt": {
            "type": "string",
            "example": "2026-09-01T13:00:00.000Z"
          },
          "endAt": {
            "type": "string",
            "example": "2026-09-01T16:00:00.000Z"
          },
          "checkinOpenAt": {
            "type": "string",
            "example": "2026-09-01T12:30:00.000Z"
          },
          "lateAt": {
            "type": "string",
            "example": "2026-09-01T14:00:00.000Z"
          },
          "checkinCloseAt": {
            "type": "string",
            "example": "2026-09-01T15:00:00.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "CANCELLED"
            ]
          },
          "timeState": {
            "type": "string",
            "enum": [
              "UPCOMING",
              "CHECKIN_OPEN",
              "ONGOING",
              "COMPLETED"
            ]
          },
          "createdBy": {
            "type": "string"
          },
          "creator": {
            "$ref": "#/components/schemas/ActivityCreatorDto"
          },
          "beaconCount": {
            "type": "number",
            "description": "จำนวนบีคอนที่ลิงก์อยู่"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-22T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-08-22T00:00:00.000Z"
          },
          "beacons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BeaconResponseDto"
            }
          },
          "attendanceSummary": {
            "$ref": "#/components/schemas/AttendanceSummaryDto"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "location",
          "startAt",
          "endAt",
          "checkinOpenAt",
          "lateAt",
          "checkinCloseAt",
          "status",
          "timeState",
          "createdBy",
          "creator",
          "beaconCount",
          "createdAt",
          "updatedAt",
          "beacons",
          "attendanceSummary"
        ]
      },
      "CreateActivityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Orientation 2026"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "กิจกรรมต้อนรับนักศึกษาใหม่"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "example": "ห้องประชุมใหญ่"
          },
          "startAt": {
            "type": "string",
            "example": "2026-09-01T13:00:00.000Z",
            "description": "เวลาเริ่มกิจกรรม (UTC)"
          },
          "endAt": {
            "type": "string",
            "example": "2026-09-01T16:00:00.000Z",
            "description": "เวลาสิ้นสุดกิจกรรม (UTC)"
          },
          "checkinOpenAt": {
            "type": "string",
            "example": "2026-09-01T12:30:00.000Z",
            "description": "เวลาเปิดเช็คชื่อ (UTC)"
          },
          "lateAt": {
            "type": "string",
            "example": "2026-09-01T14:00:00.000Z",
            "description": "เกณฑ์มาสาย — เช็คหลังเวลานี้ถือว่า LATE (UTC)"
          },
          "checkinCloseAt": {
            "type": "string",
            "example": "2026-09-01T15:00:00.000Z",
            "description": "เวลาปิดเช็คชื่อ (UTC)"
          }
        },
        "required": [
          "name",
          "startAt",
          "endAt",
          "checkinOpenAt",
          "lateAt",
          "checkinCloseAt"
        ]
      },
      "UpdateActivityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Orientation 2026"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "กิจกรรมต้อนรับนักศึกษาใหม่"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "example": "ห้องประชุมใหญ่"
          },
          "startAt": {
            "type": "string",
            "example": "2026-09-01T13:00:00.000Z",
            "description": "เวลาเริ่มกิจกรรม (UTC)"
          },
          "endAt": {
            "type": "string",
            "example": "2026-09-01T16:00:00.000Z",
            "description": "เวลาสิ้นสุดกิจกรรม (UTC)"
          },
          "checkinOpenAt": {
            "type": "string",
            "example": "2026-09-01T12:30:00.000Z",
            "description": "เวลาเปิดเช็คชื่อ (UTC)"
          },
          "lateAt": {
            "type": "string",
            "example": "2026-09-01T14:00:00.000Z",
            "description": "เกณฑ์มาสาย — เช็คหลังเวลานี้ถือว่า LATE (UTC)"
          },
          "checkinCloseAt": {
            "type": "string",
            "example": "2026-09-01T15:00:00.000Z",
            "description": "เวลาปิดเช็คชื่อ (UTC)"
          }
        }
      },
...
      "AttendanceStudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "studentCode": {
            "type": "string",
            "example": "660112230038"
          },
          "name": {
            "type": "string",
            "example": "Somchai Jaidee"
          }
        },
        "required": [
          "id",
          "studentCode",
          "name"
        ]
      },
      "AttendanceBeaconDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "CS Room 101"
          },
          "hwid": {
            "type": "string",
            "example": "32af519e88"
          }
        },
        "required": [
          "id",
          "name",
          "hwid"
        ]
      },
      "AttendanceActorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "example": "somchai"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "AttendanceRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "student": {
            "$ref": "#/components/schemas/AttendanceStudentDto"
          },
          "checkInAt": {
            "type": "string",
            "example": "2026-09-01T01:00:00.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "PRESENT",
              "LATE",
              "ABSENT",
              "EXCUSED"
            ]
          },
          "checkinMethod": {
            "type": "string",
            "enum": [
              "BEACON",
              "MANUAL"
            ]
          },
          "beacon": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AttendanceBeaconDto"
              }
            ]
          },
          "checkedInBy": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AttendanceActorDto"
              }
            ]
          },
          "manualReason": {
            "type": "object",
            "nullable": true,
            "example": "นักศึกษาปิด Bluetooth"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-09-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-09-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "student",
          "checkInAt",
          "status",
          "checkinMethod",
          "beacon",
          "checkedInBy",
          "manualReason",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "AttendanceItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "activityId": {
            "type": "string"
          },
          "activityName": {
            "type": "string",
            "example": "ปฐมนิเทศน์นักศึกษาใหม่"
          },
          "checkInAt": {
            "type": "string",
            "example": "2026-08-22T02:00:00.000Z",
            "description": "เวลาเช็คชื่อ (UTC) แสดงผลเป็น Asia/Bangkok"
          },
          "status": {
            "type": "string",
            "enum": [
              "PRESENT",
              "LATE",
              "ABSENT",
              "EXCUSED"
            ]
          },
          "checkinMethod": {
            "type": "string",
            "enum": [
              "BEACON",
              "MANUAL"
            ]
          }
        },
        "required": [
          "id",
          "activityId",
          "activityName",
          "checkInAt",
          "status",
          "checkinMethod"
        ]
      },
      "MyAttendanceDto": {
        "type": "object",
        "properties": {
          "checkInAt": {
            "type": "string",
            "example": "2026-08-22T02:00:00.000Z",
            "description": "เวลาเช็คชื่อ (UTC) แสดงผลเป็น Asia/Bangkok"
          },
          "status": {
            "type": "string",
            "enum": [
              "PRESENT",
              "LATE",
              "ABSENT",
              "EXCUSED"
            ]
          },
          "checkinMethod": {
            "type": "string",
            "enum": [
              "BEACON",
              "MANUAL"
            ]
          }
        },
        "required": [
          "checkInAt",
          "status",
          "checkinMethod"
        ]
      },
      "MyActivityItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "ปฐมนิเทศน์นักศึกษาใหม่"
          },
          "description": {
            "type": "object",
            "nullable": true,
            "example": "กิจกรรมต้อนรับนักศึกษาใหม่"
          },
          "location": {
            "type": "object",
            "nullable": true,
            "example": "ห้องประชุมใหญ่"
          },
          "startAt": {
            "type": "string",
            "example": "2026-08-22T09:00:00.000Z",
            "description": "เวลาเริ่มกิจกรรม (UTC) แสดงผลเป็น Asia/Bangkok"
          },
          "endAt": {
            "type": "string",
            "example": "2026-08-22T12:00:00.000Z",
            "description": "เวลาสิ้นสุดกิจกรรม (UTC)"
          },
          "checkinOpenAt": {
            "type": "string",
            "example": "2026-08-22T08:30:00.000Z",
            "description": "เวลาเปิดเช็คชื่อ (UTC)"
          },
          "lateAt": {
            "type": "string",
            "example": "2026-08-22T09:15:00.000Z",
            "description": "เกณฑ์มาสาย (UTC)"
          },
          "checkinCloseAt": {
            "type": "string",
            "example": "2026-08-22T11:00:00.000Z",
            "description": "เวลาปิดเช็คชื่อ (UTC)"
          },
          "timeState": {
            "type": "string",
            "enum": [
              "UPCOMING",
              "CHECKIN_OPEN",
              "ONGOING",
              "COMPLETED"
            ]
          },
          "myAttendance": {
            "nullable": true,
            "description": "การเช็คชื่อของผู้เรียกในกิจกรรมนี้ (null = ยังไม่ได้เช็ค)",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MyAttendanceDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "location",
          "startAt",
          "endAt",
          "checkinOpenAt",
          "lateAt",
          "checkinCloseAt",
          "timeState",
          "myAttendance"
        ]
      }
    }
  }
}